Implement RANLUX++ generator
RANLUX++ is an LCG equivalent of RANLUX using 576 bit numbers.

Described in
A. Sibidanov, *A revision of the subtract-with-borrow random numbergenerators*,
*Computer Physics Communications*, 221(2017), 299-303,
preprint https://arxiv.org/pdf/1705.03123.pdf

The code is loosely based on the Assembly implementation by A. Sibidanov
available at https://github.com/sibidanov/ranluxpp/.

---

This initial version of the generator corresponds to commit
https://github.com/root-project/root/commit/36cc2a820a2b5a687bb603fcaac5f2c09e574ea2
and reproduces the sequence of the assembler version.