Convert LCG state to RANLUX numbers
Just using bits from the LCG state introduces a bias because the
number is guaranteed to be smaller than m and therefore not equally
distributed over the whole range of 2 ** 576. Instead of storing the
LCG state, store the RANLUX numbers and the carry bit and convert
when needed according to notes by Lüscher. Refactor the computation
of r to a common function used by both mod_m and to_ranlux.

---

This version of the generator corresponds to commit
https://github.com/root-project/root/commit/55173bd273fad28897c29bdd261b3868a6158de4