Add compatibility engines
These engines can be used to obtain the same sequences of numbers
as RANLUX generators using recursive subtract-with-borrow steps,
but with enhanced performance. Apart from the choice of parameters,
the main difference between the various implementations is the way
of seeding the initial state of the generator.

This commit includes engines for compatibility with:
 * the original implementation by Fred James, with parameters for
   - luxury level 3 (p = 223), also matching gsl_rng_ranlux
   - luxury level 4 (p = 389), also matching gsl_rng_ranlux389
   producing floating point numbers from 24 bits of randomness;
 * the family of generators using a second-generation version of the
   RANLUX algorithm as implemented in the GNU Scientific Library:
   - gsl_rng_ranlxs[012] using 24 bits per floating point number, and
   - gsl_rng_ranlxd[12] using 48 bits per floating point number;
 * the implementation by Martin Lüscher written in C that uses four
   states per generator; similar to GSL, there are ranlxs[012] with
   24 bits per number and ranlxd[12] with 48 bits per number; and
 * the generators std::ranlux{24,48} defined by the C++ standard.

The values in the tests were extracted directly from the mentioned
implementations, showing that the LCG implementation is equivalent
to the RANLUX algorithm.

I am not adding compatibility engines for CLHEP because its semantics
are very weird: While CLHEP::RanluxEngine::setSeed yields the same
sequences as the original implementation by James, the seed is treated
differently when passed as an argument to the constructor.

---

This version corresponds to commit
https://github.com/root-project/root/commit/86f17eba38ec97153328c459f1efade18fb9eafe